0.3.0a Distribution
Version 0.3 has updates in many places, but especially in the transport
layer and in ALP/NDEF.  I am gearing up to implement the SCTP-based
datastream subsystem, and potentially also a CoAP variant, using ALP as
a universal data wrapper.

I decided to release 0.3 a bit early.  It seems like everything is
working, but my CCS went kaput and now I'm reinstalling it (this
happens every 2-3 months, actually).  What is left to test, especially,
is ALP/NDEF input.  I've compiled Demo_Opmode and Demo_PaLFI
successfully on an assortment of CC430 and MSP430 hardware, and I've
tested the basics.  But my debuggers are no longer working, which in
the past has been because CCS has become corrupted.

Happy Coding.  Expect updates tomorrow (3/8) and sunday (5/8).